fix: Failed to clean up cgroup dir when deleting Pod
The Kata shim may not always be able to migrate the QEMU
process to the parent cgroup during the MoveTo operation,
which can lead to failure when deleting the cgroup.

Therefore, we should add retry logic when deleting the cgroup
to ensure that the cgroup directory can be cleaned up properly.

Signed-off-by: wushukun <wushukun@kuaishou.com>