runtime: gpu: snp: Allow default_{vcpus,memory} annotations
Follow the same idea as in abf52420a4cd and bdbee78517e4 for snp and
coco-dev, respectively.

Allow users to set the default_vcpus and default_memory when using
nvidia-gpu-snp as they may hit issues related to the size of the
container image that must be pulled and unpacked inside the guest.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>