runtime: Make csi-directvolume-kata works well with runtime
As csi-directvolume-kata will use raw file on the host as the backend
of a virtio-blk block device, and it is treated as just a regular file,
which cause runtime label it unsupported device type. To correct such
case, this patch supports it.
Another important factor related to the issue #11296, as it said that,
the Options with bind does cause it fail. But in my test with runtime-rs,
it works well. So in this patch, I just remove the related Option
assignment from the m.Option

Fixes #11296

Signed-off-by: alex.lyn <alex.lyn@antgroup.com>