kata-types: Introduce a helper method to adjust rootfs mounts
A function `adjust_rootfs_mounts` is introduced, to manage the root
filesystem mounts based on whether a guest-pull mechanism is active.

If guest_pull is explicitly false:
- It returns the rootfs_mounts list as is, indicating no special rootfs
adjustment for guest pull, just return it.
If guest_pull is explicitly true:
- It disregards any provided rootfs_mounts. Instead, it forcefully creates
a single, default KataVirtualVolume specifically for guest-pull operations.

This volume's representation is then base64-encoded and added as the only
option to a new, singular Mount entry, which becomes the sole item in the
returned Vec<Mount>.

Signed-off-by: alex.lyn <alex.lyn@antgroup.com>