tests: Do not use `kubectl -it` for nginx-connectivity test
The following error was observed on the s390x runners:

```
Unable to use a TTY - input is not a terminal or the right kind of file
If you don't see a command prompt, try pressing enter.
```

This can happen when running in a GitHub Actions runner or other
non-interactive environment. This commit prevents a test from using
an `-it` option for kubectl for nginx-connectivity test.

Signed-off-by: Hyounggyu Choi <Hyounggyu.Choi@ibm.com>