shim: skip new MNT_NS when using cached VM
When a VM is cached, creating a new mount namespace causes
bindMountContainerRootfs to fail. This happens because the VM
was created by the factory command in other namespace
(this mnt namespace is equal to pid 1),
which then differs from the mnt namespace of containerd-shim-kata-v2.

To prevent this inconsistency,
skip creating a new mnt namespace when using a cached VM.

Fixes: #11459

Signed-off-by: Shunsuke Kimura <pbrehpuum@gmail.com>