agent: Use cgroups-rs for cgroup management
This is a follow-up patch to
https://github.com/kata-containers/kata-containers/pull/11598, aimed at
bump cgroups-rs to 0.4.1, so that the two Rust components share the same
codebase to manage cgroups.

Introduce two new types, `SandboxCgroupManager` and
`ContainerCgroupManager`. The `SandboxCgroupManager` is used to manage
sandbox resources. Device cgroups have been supported so far. The
`ContainerCgroupManager` is used to manage container resources. It has a
copy of the sandbox cgroup manager, so it can update sandbox if needed.

Signed-off-by: Xuewei Niu <niuxuewei.nxw@antgroup.com>