kata-types: Correctly handle cc_init_data with empty string
If the annotation of cc_init_data is with a empty string, it'll go
to failure with error message "failed to fill whole buffer" which is
not handled correctly within `initdata::add_hypervisor_initdata_overrides`.

To address this issue, just return Ok("") when the input is empty and
related test cases are added.

Signed-off-by: Alex Lyn <alex.lyn@antgroup.com>